sdk/python/pyproject.toml
[project]
name = "sideseat"
dynamic = ["version"]
description = "Debug, trace, and understand your AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "SideSeat", email = "support@sideseat.ai" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
    "Typing :: Typed",
]
keywords = ["opentelemetry", "observability", "llm", "ai", "tracing", "telemetry"]

dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]

[project.optional-dependencies]
# Framework-specific instrumentation
langchain = ["openinference-instrumentation-langchain>=0.1.0"]
langgraph = ["openinference-instrumentation-langchain>=0.1.0"]
crewai = ["openinference-instrumentation-crewai>=0.1.0"]
autogen = ["openinference-instrumentation-autogen-agentchat>=0.1.0"]
openai-agents = ["logfire>=0.50.0"]
pydantic-ai = ["logfire>=0.50.0"]

# Bundle all instrumentation extras
all = [
    "openinference-instrumentation-langchain>=0.1.0",
    "openinference-instrumentation-crewai>=0.1.0",
    "openinference-instrumentation-autogen-agentchat>=0.1.0",
    "logfire>=0.50.0",
]

# Development dependencies
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.urls]
Homepage = "https://sideseat.ai"
Documentation = "https://sideseat.ai/docs"
Repository = "https://github.com/sideseat/sideseat"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/sideseat/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/sideseat"]

# =============================================================================
# Testing
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]

# =============================================================================
# Ruff (formatting + linting)
# =============================================================================

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# =============================================================================
# MyPy (type checking)
# =============================================================================

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
